Rest of App Page Designs
Tab 3 is designed like the category selection screen, Tab 4 is designed like the question difficulty scren, and Tab 5 is the main question screen. Currently trying to get a back button to work in the header on Tab 5 but it isn't showing up so there is a button on the bottom in the meantime for going back.